Modularize permanent loss mod hooks
